folder/requirements.txt
mysql-connector-python==8.0.33
requests==2.31.0
Flask==2.2.5
